requirements.txt
openai-whisper
modelscope
typeguard==2.3.1
sox
bitstring
pytorch_wavelets
tensorboardX
matplotlib
numpy==1.22.4
torch
torchaudio
librosa